requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
duckdb>=0.10.0
pandas>=2.0.0
pyarrow>=14.0.0
boto3>=1.34.0
python-dotenv>=1.0.0
pyyaml>=6.0
tabulate>=0.9.0
pytest>=8.0.0
